% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_uniprot_ids.R
\name{annotate_uniprot_ids}
\alias{annotate_uniprot_ids}
\title{annotate_uniprot_ids
Function to interact with the uniprot api and get data corresponding to each
uniprot id.}
\usage{
annotate_uniprot_ids(uniprot_ids, columns = NULL, batch_size = 150)
}
\arguments{
\item{uniprot_ids}{uniprot acession numbers}

\item{columns}{This is a string  of the column names that you would like to include followed by commas.
Column names accepted by the API can be found here: https://www.uniprot.org/help/return_fields.
If not specified, the default columns are returned.}

\item{batch_size}{this is the size to include in a batch search}
}
\value{
a tibble containing requested information.
}
\description{
annotate_uniprot_ids
Function to interact with the uniprot api and get data corresponding to each
uniprot id.
}
\examples{
annotate_uniprot_ids("P01308")
}
