% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fasta_info.R
\name{extract_fasta_info}
\alias{extract_fasta_info}
\title{extract fasta info}
\usage{
extract_fasta_info(fasta_file)
}
\arguments{
\item{fasta_file}{}
}
\value{
a data frame with organism name, organism id and sequence
}
\description{
extracts organism name, organism id and sequence from fasta file
}
