% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_reference_proteomes.R
\name{get_all_reference_proteomes}
\alias{get_all_reference_proteomes}
\title{get_all_reference_proteomes}
\usage{
get_all_reference_proteomes()
}
\value{
a tibble
}
\description{
This function returns all current reference proteomes from uniprot. This
gets updated every 8 weeks per their documentation. See
https://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/reference_proteomes/README
for more info.
}
\details{
Contains the following columns:
Proteome_ID,Tax_ID,OSCODE,SUPERREGNUM,#(1),#(2),#(3),Species Name
#(1),#(2),#(3) are statistics as follows:
}
\examples{
rps <- get_all_reference_proteomes()
}
