% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotations_from_uniprot.R
\name{get_annotations_from_uniprot}
\alias{get_annotations_from_uniprot}
\title{get_go_kegg_eggnogg_annotation}
\usage{
get_annotations_from_uniprot(uniprot_ids)
}
\arguments{
\item{pg_matrix_fp}{pg_matrix output of diann}

\item{outdir}{out directory}
}
\value{
tsv at output directory location
}
\description{
given the pg_matrix output of diann, assign go terms to each protein group
and summarise.
}
