% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eggnog_function.R
\name{get_eggnog_function}
\alias{get_eggnog_function}
\title{get_eggnog_function}
\usage{
get_eggnog_function(eggnog_ids)
}
\arguments{
\item{eggnog_ids}{}
}
\description{
Get Eggnog functional go terms given a vector of EGGNOG ids
}
