% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_from_organism_ids.R
\name{get_fasta_from_organism_ids}
\alias{get_fasta_from_organism_ids}
\title{create_fasta_from_organism_ids}
\usage{
get_fasta_from_organism_ids(
  organism_ids,
  destination_fp = paste0(getwd(), "/", Sys.Date(), ".fasta"),
  additonal_organism_ids = NULL
)
}
\arguments{
\item{organism_ids}{vector of organism ids}

\item{destination_fp}{path to the destination fasta file. Must be .fasta}

\item{additonal_organism_ids}{vector of any additional organism ids to include}
}
\value{
a downloaded .fasta file to user specified file path.
}
\description{
Given a vector of organism ids, pull down a list of proteomes from uniprot
for each organism. The first one from the list is then downloaded as a fasta
file. This is then repeated for each organism id and concatintated into a
combined fasta file.
}
\details{
This is particularly usefull for generating a metaproteomics .fasta database
}
\examples{
# Making a database with Human and E.coli
create_fasta_from_organism_ids(organism_ids = c(9606, 562))
}
