% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kegg_in_batches.R
\name{get_kegg_in_batches}
\alias{get_kegg_in_batches}
\title{get_kegg_in_batches}
\usage{
get_kegg_in_batches(kegg_ids, batch_size = 10)
}
\arguments{
\item{kegg_ids}{= kegg ides in the format "hsa:1234"}
}
\description{
Given a vector of kegg ids, get all of the functional information using the
KEGG API.
}
