% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_taxonomy.R
\name{get_ncbi_taxonomy}
\alias{get_ncbi_taxonomy}
\title{get_ncbi_taxonomy}
\usage{
get_ncbi_taxonomy(ncbi_ids)
}
\arguments{
\item{ncbi_ids}{a vector of ncbi taxonomy ids}
}
\description{
given a list of ncbi taxonomy, fetch complete taxonomic information.
}
