% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_limma_analysis.R
\name{perform_limma_analysis}
\alias{perform_limma_analysis}
\title{perform_limma_analysis}
\usage{
perform_limma_analysis(
  qf_object,
  assay_name,
  design_df,
  formula,
  contrast,
  normalize = TRUE
)
}
\arguments{
\item{qf_object}{qfeatures object}

\item{assay_name}{assay name}

\item{design_df}{design_data frame}

\item{formula}{formula}

\item{contrast}{contrasts to look for}

\item{normalize}{whether to log2 normalise}
}
\description{
Allows user to perfom limma analysis on any "assay" in the QFeatures object
}
