% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_qfeature.R
\name{prepare_qfeature}
\alias{prepare_qfeature}
\title{prepare_qfeature}
\usage{
prepare_qfeature(
  pg_matrix_fp,
  pr_matrix_fp,
  superkingdom_matrix_fp,
  kingdom_matrix_fp,
  phylum_matrix_fp,
  class_matrix_fp,
  order_matrix_fp,
  family_matrix_fp,
  genus_matrix_fp,
  species_matrix_fp,
  go_matrix_fp
)
}
\arguments{
\item{go_matrix_fp}{}
}
\description{
assemble qfeature object from all of the matricies generated by the pipeline
}
