% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_taxonomy_matricies.R
\name{prepare_taxonomy_matricies}
\alias{prepare_taxonomy_matricies}
\title{prepare_taxonomy_matricies}
\usage{
prepare_taxonomy_matricies(
  full_taxonomy_fp,
  pr_group_matrix_fp,
  fasta_fp,
  output_dir = getwd()
)
}
\arguments{
\item{full_taxonomy_fp}{file path to taxonomy file}

\item{pr_group_matrix_fp}{file path to pr_group matrix from DIAnn}

\item{fasta_fp}{file path to fasta used for experiment.}
}
\value{
taxonomy_matrix.tsv for each level of taxonomy
}
\description{
prepare_taxonomy_matricies
}
